+++ packages/cli/tsup.config.ts
@@ -3,7 +3,15 @@ import {defineConfig} from 'tsup';
 export default defineConfig({
   clean: true,
   entry: ['src/bin.ts'],
-  external: ['@dynobox/sdk', 'commander', 'execa', 'tinyglobby', 'tsx', 'yaml'],
+  external: [
+    '@dynobox/sdk',
+    'commander',
+    'execa',
+    'mockttp',
+    'tinyglobby',
+    'tsx',
+    'yaml',
+  ],
   format: ['esm'],
   noExternal: ['@dynobox/runner-local', '@dynobox/evaluators'],
   platform: 'node',